+++ examples/line_sender_cpp_example_array.cpp
@@ -16,8 +16,8 @@ static bool array_example(std::string_view host, std::string_view port)
         const auto symbol_col = "symbol"_cn;
         const auto book_col = "order_book"_cn;
         size_t rank = 3;
-        std::vector<uint32_t> shape{2, 3, 2};
-        std::vector<int32_t> strides{48, 16, 8};
+        std::vector<uintptr_t> shape{2, 3, 2};
+        std::vector<intptr_t> strides{48, 16, 8};
         std::array<double, 12> arr_data = {
             48123.5,
             2.4,
